insureops-ai/agents/requirements.txt
langchain>=0.3.0
langgraph>=0.2.0
langchain-community>=0.3.0
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.31.0
pydantic>=2.5.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.0
PyMuPDF>=1.23.0
numpy>=1.24.0
Pillow>=10.0.0
pytesseract>=0.3.10
